+++ gradio/components/json_component.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import json
 import warnings
 from collections.abc import Callable, Sequence
 from typing import (
@@ -20,7 +19,7 @@ from gradio.data_classes import JsonData
 from gradio.events import Events
 from gradio.exceptions import Error
 from gradio.i18n import I18nData
-from gradio.utils import set_default_buttons
+from gradio.utils import parse_escaped_json, set_default_buttons
 
 if TYPE_CHECKING:
     from gradio.components import Timer
@@ -169,7 +168,7 @@ class JSON(Component):
         return {"foo": "bar"}
 
     def read_from_flag(self, payload: Any):
-        return json.loads(payload)
+        return parse_escaped_json(payload)
 
     def api_info(self) -> dict[str, Any]:
         return {"type": {}, "description": "any valid json"}