+++ gradio/state_holder.py
@@ -84,7 +84,10 @@ class SessionState:
         block = self.blocks_config.blocks[key]
         if block.stateful:
             if key not in self.state_data:
-                self.state_data[key] = deepcopy(getattr(block, "value", None))
+                value = getattr(block, "value", None)
+                if callable(value):
+                    value = value()
+                self.state_data[key] = deepcopy(value)
             return self.state_data[key]
         else:
             return block