+++ js/code/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/code
 
+## 0.18.1
+
+### Dependency updates
+
+- @gradio/statustracker@0.15.1
+- @gradio/icons@0.16.0
+- @gradio/atoms@0.26.0
+- @gradio/upload@0.18.1
+
 ## 0.18.0
 
 ### Features