+++ js/paramviewer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/paramviewer
 
+## 0.10.1
+
+### Dependency updates
+
+- @gradio/statustracker@0.15.1
+- @gradio/atoms@0.26.0
+
 ## 0.10.0
 
 ### Features